# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Jesus Antona",
    "description": "Ingenier\u00eda Web",
    "version": "0.1.0"
  },
  "paths": {
...
    "/Usuarios/": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Find All Users",
        "description": "Devuelve una lista de usuarios filtrados",
        "operationId": "find_all_users_Usuarios__get",
        "parameters": [
          {
            "name": "letraNombre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Primera letra del nombre",
              "title": "Letranombre"
            },
            "description": "Primera letra del nombre"
          },
          {
            "name": "letraApellido",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Primera letra del apellido",
              "title": "Letraapellido"
            },
            "description": "Primera letra del apellido"
          },
          {
            "name": "valoracion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Valoracion del usuario",
              "title": "Valoracion"
            },
            "description": "Valoracion del usuario"
          },
          {
            "name": "esAdmin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Rol del usuario",
              "title": "Esadmin"
            },
            "description": "Rol del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario-Output"
                  },
                  "title": "Response Find All Users Usuarios  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Create User",
        "description": "Crea un Usuario y lo devuelve",
        "operationId": "create_user_Usuarios__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Find One User",
        "description": "Devuelve el usuario con id pasado por par\u00e1metro",
        "operationId": "find_one_user_Usuarios__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del usuario a buscar",
              "title": "Id"
            },
            "description": "Id del usuario a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Delete User",
        "description": "Elimina el usuario con id pasado por par\u00e1metro",
        "operationId": "delete_user_Usuarios__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del usuario a eliminar",
              "title": "Id"
            },
            "description": "Id del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Update User",
        "description": "Actualiza el usuario con id pasado por par\u00e1metro y lo devuelve",
        "operationId": "update_user_Usuarios__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del usuario a actualizar",
              "title": "Id"
            },
            "description": "Id del usuario a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Usuario-Input"
                  }
                ],
                "description": "Datos del usuario a actualizar",
                "title": "User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Usuarios/Username/{username}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Find User By Username",
        "description": "Devuelve el usuario con username pasado por par\u00e1metro",
        "operationId": "find_User_By_Username_Usuarios_Username__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Usuario del usuario a buscar",
              "title": "Username"
            },
            "description": "Usuario del usuario a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Usuario-Input": {
        "properties": {
          "idUser": {
            "type": "string",
            "title": "Iduser"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido1": {
            "type": "string",
            "title": "Apellido1"
          },
          "apellido2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apellido2"
          },
          "valoracion": {
            "type": "number",
            "title": "Valoracion"
          },
          "esAdmin": {
            "type": "boolean",
            "title": "Esadmin"
          }
        },
        "type": "object",
        "required": [
          "idUser",
          "usuario",
          "nombre",
          "apellido1",
          "valoracion",
          "esAdmin"
        ],
        "title": "Usuario"
      },
      "Usuario-Output": {
        "properties": {
          "idUser": {
            "type": "string",
            "title": "Iduser"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido1": {
            "type": "string",
            "title": "Apellido1"
          },
          "apellido2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apellido2"
          },
          "valoracion": {
            "type": "number",
            "title": "Valoracion"
          },
          "esAdmin": {
            "type": "boolean",
            "title": "Esadmin"
          }
        },
        "type": "object",
        "required": [
          "idUser",
          "usuario",
          "nombre",
          "apellido1",
          "apellido2",
          "valoracion",
          "esAdmin"
        ],
        "title": "Usuario"
      },
      "UsuarioCreate": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido1": {
            "type": "string",
            "title": "Apellido1"
          },
          "apellido2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apellido2"
          },
          "valoracion": {
            "type": "number",
            "title": "Valoracion"
          },
          "esAdmin": {
            "type": "boolean",
            "title": "Esadmin"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "nombre",
          "apellido1",
          "valoracion",
          "esAdmin"
        ],
        "title": "UsuarioCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ]
}